dashboard/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
boto3==1.34.0
pandas==2.0.3
python-dateutil==2.8.2

